separate pytest fixtures
