+++ conftest.py
@@ -1,0 +1,5 @@
+pytest_plugins = [
+    "gyandex.podgen.storage.test_fixtures",
+    "gyandex.podgen.feed.test_fixtures",
+    "gyandex.podgen.engine.test_fixtures",
+]
